+++ pyfor/__init__.py
@@ -1,12 +1,12 @@
 from __future__ import absolute_import
 
-__version__ = '0.2'
+__version__ = '0.3'
 
 from pyfor import cloud
 from pyfor import rasterizer
 from pyfor import gisexport
 from pyfor import clip_funcs
-from pyfor import filter
+from pyfor import ground_filter
 from pyfor import plot
 from pyfor import collection
 from pyfor import voxelizer